Fix MSVC warnings (#55)
* Fix MSVC C++ warnings

* Fix link.exe warning

* PR feedback, fix x64 pointer truncation warning

+++ src/win32_io.c
@@ -41,5 +41,5 @@ int CreateTempFile (char *filename)
 		return FileNotFound;
 	}
 
-	return (int)fopen (filename, "w");
+	return (int)(intptr_t)fopen (filename, "w");
 }
